+++ flash/core/data/input_transform.py
@@ -865,8 +865,10 @@ class InputTransform(Properties):
 
     def _get_current_transform(self, process_state: ProcessState):
         fn = self.get_state(process_state)
-        if fn is not None and fn.transform is not None:
-            return fn.transform
+        if fn is not None:
+            if fn.transform is not None:
+                return fn.transform
+            return self._identity
         return self.current_transform
 
     @partial(transform_context, current_fn="per_sample_transform_on_device")
