+++ flash/image/keypoint_detection/cli.py
@@ -68,7 +68,6 @@ def keypoint_detection():
             "model.num_keypoints": 1,
             "trainer.max_epochs": 3,
         },
-        legacy=True,
     )
 
     cli.trainer.save_checkpoint("keypoint_detection_model.pt")
